requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pydantic-settings==2.2.1

openai==1.30.1
google-generativeai==0.8.3
tiktoken==0.7.0

networkx==3.2.1
graspologic==3.3.0
numpy==1.26.4
scikit-learn==1.4.2

faiss-cpu==1.8.0

datasets==2.19.1

httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3